refactor : content 컴포넌트
